@@ -21,10 +21,8 @@ public class Protocol implements View.OnLongClickListener {
     public final int POWER2 = 0b110;
     public final int X_SQR_Y2 = 0b111;
 
-    Context context;
     private TextView tvProtocol;
     private FragmentManager manager;
-    //    private AlertDialog.Builder builder;
     private MyDialogFragment myDialogFragment;
 
     private boolean doCls;
@@ -33,27 +31,6 @@ public class Protocol implements View.OnLongClickListener {
 
     private ComplexFormat complexFormat;
     Mode mode;
-
-//    public Protocol(Context context, TextView tvProtocol, Mode mode) {
-//        this.context = context;
-//        this.tvProtocol = tvProtocol;
-//        this.mode = mode;
-//        tvProtocol.setOnLongClickListener(this);
-//        sbProtocol = new StringBuilder("");
-//        complexFormat = new ComplexFormat();
-//        doCls = true;
-//    }
-//
-//    public Protocol(MainActivity context, TextView tvProtocol, Mode mode, View adv) {
-//        this.context = context;
-//        this.tvProtocol = tvProtocol;
-//        this.adv = adv;
-//        this.mode = mode;
-//        tvProtocol.setOnLongClickListener(this);
-//        sbProtocol = new StringBuilder("");
-//        complexFormat = new ComplexFormat();
-//        doCls = true;
-//    }
 
     Protocol(TextView tvProtocol, Mode mode, MyDialogFragment myDialogFragment, FragmentManager manager) {
         this.tvProtocol = tvProtocol;
@@ -147,16 +124,16 @@ public class Protocol implements View.OnLongClickListener {
         sbProtocol.append("(").append(complexFormat.format(complex)).append(")\n");
     }
 
-    void println(List cplxNumList) {
+    void println(List<Complex> cplxNumList) {
         if (doCls) {
             cls();
         }
-        Complex complex = (Complex) cplxNumList.get(0);
+        Complex complex = cplxNumList.get(0);
         tvProtocol.append("(" + complexFormat.format(complex) + "), ...\n");
         doCls = false;
 
         for (int i = 0; i < cplxNumList.size(); i++) {
-            complex = (Complex) cplxNumList.get(i);
+            complex = cplxNumList.get(i);
             sbProtocol.append("\n(").append(complexFormat.format(complex)).append(")");
         }
         doCls = true;
@@ -184,7 +161,6 @@ public class Protocol implements View.OnLongClickListener {
     }
 
     void print(String msg) {
-        //protocol.jTextArea1.append(msg);
         if (doCls) {
             cls();
         }
@@ -251,16 +227,16 @@ public class Protocol implements View.OnLongClickListener {
         sbProtocol.append("(").append(complexFormat.format(complex)).append(")");
     }
 
-    void print(List cplxNumList) {
+    void print(List<Complex> cplxNumList) {
         if (doCls) {
             cls();
         }
-        Complex complex = (Complex) cplxNumList.get(0);
+        Complex complex = cplxNumList.get(0);
         tvProtocol.append("(" + complexFormat.format(complex) + ")");
         doCls = false;
 
         for (int i = 0; i < cplxNumList.size(); i++) {
-            complex = (Complex) cplxNumList.get(i);
+            complex = cplxNumList.get(i);
             sbProtocol.append("\n+(").append(complexFormat.format(complex)).append(")");
         }
 
